requirements.txt
plyer
kivymd_extensions.sweetalert
kaki
